rewrite reload code
